+++ cppwamp/include/cppwamp/codec.hpp
@@ -39,56 +39,112 @@ struct CPPWAMP_API KnownCodecIds
 
 
 //******************************************************************************
+// Output Sinks
+//******************************************************************************
+
+//------------------------------------------------------------------------------
+/** This primary template is specialized for each type that can be the
+    output of a codec encoding operation. */
+//------------------------------------------------------------------------------
+template <typename TOutput, typename TEnabled = void>
+class OutputSink
+{};
+
+//------------------------------------------------------------------------------
+/** Output sink specialization referencing a std::string. */
+//------------------------------------------------------------------------------
+template <>
+class OutputSink<std::string>
+{
+public:
+    using Output = std::string;
+    OutputSink(Output& s) : output_(&s) {}
+    Output& output() const {return *output_;}
+
+private:
+    Output* output_;
+};
+
+//------------------------------------------------------------------------------
+/** Output sink specialization referencing a wamp::MessageBuffer. */
+//------------------------------------------------------------------------------
+template <>
+class OutputSink<MessageBuffer>
+{
+public:
+    using Output = MessageBuffer;
+    OutputSink(Output& b) : output_(&b) {}
+    Output& output() const {return *output_;}
+
+private:
+    Output* output_;
+};
+
+//------------------------------------------------------------------------------
+/** Output sink specialization referencing a std::ostream. */
+//------------------------------------------------------------------------------
+template <>
+class OutputSink<std::ostream>
+{
+public:
+    using Output = std::ostream;
+    OutputSink(Output& b) : output_(&b) {}
+    Output& output() const {return *output_;}
+
+private:
+    Output* output_;
+};
+
+//------------------------------------------------------------------------------
+using StringSink = OutputSink<std::string>;   ///< Output sink for strings.
+using BufferSink = OutputSink<MessageBuffer>; ///< Output sink for message buffers.
+using StreamSink = OutputSink<std::ostream>;  ///< Output sink for streams.
+
+//------------------------------------------------------------------------------
+template <typename TOutput, typename TEnabled = void>
+struct OutputTraits {};
+
+template <>
+struct OutputTraits<std::string> { using Sink = StringSink; };
+
+template <>
+struct OutputTraits<MessageBuffer> { using Sink = BufferSink; };
+
+template <typename S>
+struct OutputTraits<S, EnableIf<std::is_base_of<std::ostream, S>::value, void>>
+{
+    using Sink = StreamSink;
+};
+
+//------------------------------------------------------------------------------
+/** Yields the OutputSink specialization associated with the given output type. */
+//------------------------------------------------------------------------------
+template <typename TOutput>
+using SinkTypeFor = typename OutputTraits<ValueTypeOf<TOutput>>::Sink;
+
+
+//******************************************************************************
 // Encoder
 //******************************************************************************
 
-/** Output category for containers of bytes which provide `push_back` and
-    `insert` member functions. */
-struct ByteContainerOutputCategory {};
 
-/** Output category for output streams of bytes. */
-struct StreamOutputCategory {};
+//------------------------------------------------------------------------------
+/** Primary template specialized for codec format tag and sink combinations . */
+//------------------------------------------------------------------------------
+template <typename TFormat, typename TSink, typename TEnable = void>
+class SinkEncoder {};
 
-/** Type used to indicate output category detection failed. */
-struct UnknownOutputCategory {};
-
-/** Traits class that determines the category type that best matches the
-    given output type. */
-template <typename O, typename Enabled = void>
-struct OutputCategory
-{
-    using type = UnknownOutputCategory;
-};
-
-template <typename O>
-struct OutputCategory<O, EnableIf<std::is_base_of<std::ostream, O>::value,
-                                  void>>
-{
-    using type = StreamOutputCategory;
-};
-
-template <typename O>
-struct OutputCategory<O, EnableIf<!std::is_base_of<std::ostream, O>::value,
-                                  void>>
-{
-    using type = ByteContainerOutputCategory;
-};
-
-/// Yields the category type that best matches the given output type.
-template <typename O>
-using OutputCategoryTypeOf = typename OutputCategory<ValueTypeOf<O>>::type;
 
 //------------------------------------------------------------------------------
 /** Yields the encoder type needed to encode a Variant to the given output type
-    and output category
+    using the given format.
     @tparam F Encoder format type tag (e.g. Json)
     @tparam O Output type (e.g. std::string)
-    @tparam C Output category (deduced)
     @see wamp::Decoder
     @see wamp::encode */
 //------------------------------------------------------------------------------
-template <typename F, typename O, typename C = OutputCategoryTypeOf<O>>
-using Encoder = typename F::template Encoder<ValueTypeOf<O>, C>;
+template <typename F, typename O>
+using Encoder = SinkEncoder<F, SinkTypeFor<O>>;
 
 //------------------------------------------------------------------------------
 /** Encodes the given variant to the given byte container or stream.
@@ -108,56 +164,110 @@ void encode(const Variant& variant, TOutput&& output)
 
 
 //******************************************************************************
+// Input Sources
+//******************************************************************************
+
+//------------------------------------------------------------------------------
+/** This primary template is specialized for each type that can be the
+    input of a codec decoding operation. */
+//------------------------------------------------------------------------------
+template <typename TInput, typename TEnabled = void>
+class InputSource
+{};
+
+//------------------------------------------------------------------------------
+/** Input source specialization referencing a std::string. */
+//------------------------------------------------------------------------------
+template <>
+class InputSource<std::string>
+{
+public:
+    using Input = std::string;
+    InputSource(const Input& s) : input_(&s) {}
+    const Input& input() const {return *input_;}
+
+private:
+    const Input* input_;
+};
+
+//------------------------------------------------------------------------------
+/** Input source specialization referencing a wamp::MessageBuffer. */
+//------------------------------------------------------------------------------
+template <>
+class InputSource<MessageBuffer>
+{
+public:
+    using Input = MessageBuffer;
+    InputSource(const Input& b) : input_(&b) {}
+    const Input& input() const {return *input_;}
+
+private:
+    const Input* input_;
+};
+
+//------------------------------------------------------------------------------
+/** Input source specialization referencing a std::ostream. */
+//------------------------------------------------------------------------------
+template <>
+class InputSource<std::istream>
+{
+public:
+    using Input = std::istream;
+    InputSource(Input& b) : input_(&b) {}
+    Input& input() const {return *input_;}
+
+private:
+    Input* input_;
+};
+
+//------------------------------------------------------------------------------
+using StringSource = InputSource<std::string>;   ///< Input source for strings.
+using BufferSource = InputSource<MessageBuffer>; ///< Input source for message buffers.
+using StreamSource = InputSource<std::istream>;  ///< Input source for streams.
+
+//------------------------------------------------------------------------------
+template <typename TInput, typename TEnabled = void>
+struct InputTraits {};
+
+template <>
+struct InputTraits<std::string> { using Source = StringSource; };
+
+template <>
+struct InputTraits<MessageBuffer> { using Source = BufferSource; };
+
+template <typename S>
+struct InputTraits<S, EnableIf<std::is_base_of<std::istream, S>::value, void>>
+{
+    using Source = StreamSource;
+};
+
+//------------------------------------------------------------------------------
+/** Yields the InputSource specialization associated with the given input type. */
+//------------------------------------------------------------------------------
+template <typename TInput>
+using SourceTypeFor = typename InputTraits<ValueTypeOf<TInput>>::Source;
+
+
+//******************************************************************************
 // Decoder
 //******************************************************************************
 
-/** Input category for contiguous byte arrays which provide `data` and `size`
-    member functions. */
-struct ByteArrayInputCategory {};
-
-/** Input category for input streams of bytes. */
-struct StreamInputCategory {};
-
-/** Type used to indicate input category detection failed. */
-struct UnknownInputCategory {};
-
-/** Traits class that determines the category type that best matches the
-    given input type. */
-template <typename I, typename Enabled = void>
-struct InputCategory
-{
-    using type = UnknownOutputCategory;
-};
-
-template <typename I>
-struct InputCategory<I, EnableIf<std::is_base_of<std::istream, I>::value,
-                                 void>>
-{
-    using type = StreamInputCategory;
-};
-
-template <typename I>
-struct InputCategory<I, EnableIf<!std::is_base_of<std::istream, I>::value,
-                                 void>>
-{
-    using type = ByteArrayInputCategory;
-};
-
-/// Yields the category type that best matches the given input type.
-template <typename I>
-using InputCategoryTypeOf = typename InputCategory<ValueTypeOf<I>>::type;
+//------------------------------------------------------------------------------
+/** Primary template specialized for codec format tag and source combinations . */
+//------------------------------------------------------------------------------
+template <typename TFormat, typename TSource, typename TEnable = void>
+class SourceDecoder {};
 
 //------------------------------------------------------------------------------
 /** Yields the decoder type needed to decode a Variant from the given input type
-    and input category
+    using the given format.
     @tparam F Encoder format type tag (e.g. Json)
     @tparam I Input type (e.g. std::string)
-    @tparam I Input category (deduced)
     @see wamp::Encoder
     @see wamp::decode */
 //------------------------------------------------------------------------------
-template <typename F, typename I, typename C = InputCategoryTypeOf<I>>
-using Decoder = typename F::template Decoder<ValueTypeOf<I>, C>;
+template <typename F, typename I>
+using Decoder = SourceDecoder<F, SourceTypeFor<I>>;;
 
 //------------------------------------------------------------------------------
 /** Decodes from the given byte sequence or stream to the given variant.